move parseResults() to request-action service
