+++ src/app/modules/group/http-services/request-actions.service.ts
@@ -4,6 +4,7 @@ import { ActionResponse, successData, objectToMap } from 'src/app/shared/http-se
 import { map } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 import { appConfig } from 'src/app/shared/helpers/config';
+import { Result } from '../helpers/response-toast';
 
 export enum Action {
   Accept,
@@ -54,4 +55,15 @@ export class RequestActionsService {
         map(objectToMap)
       );
   }
+}
+
+export function parseResults(data: Map<string, any>[]): Result {
+  const res : Result = { countRequests: 0, countSuccess: 0 };
+  data.forEach(elm => {
+    res.countRequests += elm.size;
+    res.countSuccess += Array.from(elm.values())
+      .map<number>(state => ([ 'success', 'unchanged' ].includes(state) ? 1 : 0))
+      .reduce((acc, res) => acc + res, 0);
+  });
+  return res;
 }
